requirements.txt
torch>=1.0
onnx>=1.6
opencv-python>=3.4
numpy>=1.17
pillow==7.2.0
grpcio==1.31.0
grpcio-tools==1.31.0